Clean up mocks controllers/data structure
