+++ mocks/api/controllers/dokumente.js
@@ -1,197 +1,4 @@
-const mockedDokumente = [
-  {
-    id: "3cf0f6be-5e99-11f1-9bd2-325096b39f47",
-    verfahren_id: "019c5736-7834-7ee6-99de-41d9605697d2",
-    dokumente: [
-      {
-        id: "f7e400a6-5ec1-11f1-8acd-325096b39f47",
-        einreichung_id: "3cf0f6be-5e99-11f1-9bd2-325096b39f47",
-        status: "ERSTELLT",
-        name: "Klageschrift.pdf",
-        size_in_bytes: 3211,
-        type: "XJUSTIZ",
-        viren_scan_status: "SAUBER",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-    ],
-  },
-  {
-    id: "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-    verfahren_id: "019c5741-7287-760b-865e-541a69c2aa8d",
-    dokumente: [
-      {
-        id: "71ed8f80-5f4d-11f1-8fd7-325096b39f47",
-        einreichung_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-        status: "EINGEREICHT",
-        name: "Testdatei",
-        size_in_bytes: 0,
-        type: "ANHANG",
-        viren_scan_status: "SAUBER",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-      {
-        id: "77b6ec7c-5f4d-11f1-b5b4-325096b39f47",
-        einreichung_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-        status: "VALIDIERT",
-        name: "Rückmeldung zur Angelegenheit",
-        size_in_bytes: 0,
-        type: "SCHRIFTSTUECK",
-        viren_scan_status: "INFIZIERT",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-      {
-        id: "7affe564-5f4d-11f1-8bab-325096b39f47",
-        einreichung_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6",
-        status: "EINGEREICHT",
-        name: "Word.docx (final)",
-        size_in_bytes: 0,
-        type: "SCHRIFTSTUECK",
-        viren_scan_status: "FEHLGESCHLAGEN",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-    ],
-  },
-  {
-    id: "f673a416-5e5c-11f1-8b35-325096b39f47",
-    verfahren_id: "019c5741-7287-760b-865e-541a69c2aa8d",
-    dokumente: [
-      {
-        id: "99b37434-5f4e-11f1-b909-325096b39f47",
-        einreichung_id: "f673a416-5e5c-11f1-8b35-325096b39f47",
-        status: "VALIDIERT",
-        name: "Klageeinreichung XYZ",
-        size_in_bytes: 0,
-        type: "ANHANG",
-        viren_scan_status: "IN_BEARBEITUNG",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-      {
-        id: "9620256a-5f4e-11f1-923b-325096b39f47",
-        einreichung_id: "f673a416-5e5c-11f1-8b35-325096b39f47",
-        status: "EINGEREICHT",
-        name: "Dokumentenname",
-        size_in_bytes: 0,
-        type: "SCHRIFTSTUECK",
-        viren_scan_status: "INFIZIERT",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-    ],
-  },
-  {
-    id: "4aec2e8c-5e99-11f1-abc9-325096b39f47",
-    verfahren_id: "019c5741-7287-760b-865e-541a69c2aa8d",
-    dokumente: [
-      {
-        id: "b689d206-5f4e-11f1-b0d9-325096b39f47",
-        einreichung_id: "4aec2e8c-5e99-11f1-abc9-325096b39f47",
-        status: "ERSTELLT",
-        name: "Ein weiterer Name",
-        size_in_bytes: 0,
-        type: "ANHANG",
-        viren_scan_status: "AUSSTEHEND",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-      {
-        id: "b9e4db76-5f4e-11f1-a1ee-325096b39f47",
-        einreichung_id: "4aec2e8c-5e99-11f1-abc9-325096b39f47",
-        status: "EINGEREICHT",
-        name: "Und noch ein Name",
-        size_in_bytes: 0,
-        type: "SCHRIFTSTUECK",
-        viren_scan_status: "INFIZIERT",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-    ],
-  },
-  {
-    id: "554e8956-5e99-11f1-8a89-325096b39f47",
-    verfahren_id: "019c5733-2edb-7702-84d7-77f416b5f16c",
-    dokumente: [
-      {
-        id: "a5a007d6-5f4d-11f1-b73f-325096b39f47",
-        einreichung_id: "554e8956-5e99-11f1-8a89-325096b39f47",
-        status: "EINGEREICHT",
-        name: "XJustiz Test",
-        size_in_bytes: 0,
-        type: "XJUSTIZ",
-        viren_scan_status: "SAUBER",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-      {
-        id: "a5a007d6-5f4d-11f1-b73f-325096b39f47",
-        einreichung_id: "554e8956-5e99-11f1-8a89-325096b39f47",
-        status: "EINGEREICHT",
-        name: "Name, name, name",
-        size_in_bytes: 0,
-        type: "SCHRIFTSTUECK",
-        viren_scan_status: "SAUBER",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-    ],
-  },
-  {
-    id: "5d118abc-5e99-11f1-8c77-325096b39f47",
-    verfahren_id: "019c5733-2edb-7702-84d7-77f416b5f16c",
-    dokumente: [
-      {
-        id: "da1c0ff4-5f4e-11f1-a50c-325096b39f47",
-        einreichung_id: "5d118abc-5e99-11f1-8c77-325096b39f47",
-        status: "EINGEREICHT",
-        name: "string",
-        size_in_bytes: 0,
-        type: "ANHANG",
-        viren_scan_status: "SAUBER",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-      {
-        id: "dcc3a4ba-5f4e-11f1-9c8f-325096b39f47",
-        einreichung_id: "5d118abc-5e99-11f1-8c77-325096b39f47",
-        status: "EINGEREICHT",
-        name: "Klageeinreichung.pdf",
-        size_in_bytes: 0,
-        type: "SCHRIFTSTUECK",
-        viren_scan_status: "SAUBER",
-        gesendet_am: "2026-06-02T19:57:54.675Z",
-        eingereicht_am: "2026-06-02T19:57:54.675Z",
-        erstellt_von: "string",
-        erstellt_am: "2026-06-02T19:57:54.675Z",
-      },
-    ],
-  },
-];
+import { mockedDokumente } from "../data/dokumente.js";
 
 export const getAllDokumenteByIds = (verfahrenId, einreichungId) => {
   const dokumente = mockedDokumente.find(
